WishperAI/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
av==10.0.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blessed==1.20.0
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
cmaes==0.9.1
cmake==3.26.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.0.7
croniter==1.3.15
cryptography==40.0.2
ctranslate2==3.14.0
cycler==0.11.0
datasets==2.12.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
dill==0.3.6
dm-tree==0.1.8
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
einops==0.6.1
email-validator==2.0.0.post2
etils==1.3.0
fastapi==0.88.0
faster-whisper==0.6.0
ffmpeg-python==0.2.0
filelock==3.12.0
flatbuffers==23.5.26
flax==0.6.10
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.14.1
humanfriendly==10.0
HyperPyYAML==1.2.0
idna==3.4
importlib-resources==5.12.0
inquirer==3.1.3
install==1.3.5
itsdangerous==2.1.2
jax==0.4.10
jaxlib==0.4.10+cuda11.cudnn86
Jinja2==3.1.2
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
lit==16.0.5
llvmlite==0.40.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.1.0
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
onnxruntime==1.15.0
openai-whisper==20230314
opt-einsum==3.3.0
optax==0.1.5
optuna==3.1.1
orbax-checkpoint==0.2.3
ordered-set==4.1.0
orjson==3.8.12
packaging==23.1
pandas==2.0.1
passlib==1.7.4
Pillow==9.5.0
pooch==1.6.0
primePy==1.3
protobuf==3.20.3
psutil==5.9.5
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio@11b56a137a578db9335efc00298f6ec1932e6317
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyarrow==12.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytorch-metric-learning==2.1.2
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
regex==2023.5.5
requests==2.30.0
responses==0.18.0
rich==13.3.5
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
semver==3.0.0
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
speechbrain==0.5.14
SQLAlchemy==2.0.15
starlette==0.22.0
starsessions==1.3.0
streaming-form-data==1.12.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6
tensorstore==0.1.36
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
whisper-jax @ git+https://github.com/sanchit-gandhi/whisper-jax.git@0d3bc54c8c673d71732ba6bc20fa0323b1101fa6
whisperx @ git+https://github.com/m-bain/whisperx.git@1c528d1a3cdf0ce69bfad1edc49e2e0c2bcf1d0a
xxhash==3.2.0
yarl==1.9.2
